third_party/blink/web_tests/ppapi/plugins/multiple-plugins-expected.txt

CONSOLE MESSAGE: Blink Deprecated Test Plugin: Destroying
CONSOLE MESSAGE: Blink Deprecated Test Plugin: Destroying
Test that we can create two plugins, and independently clean each.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testObj1.testObject is defined.
PASS testObj2.testObject is defined.
PASS testObj1.testObject threw exception ReferenceError: Property testObject does not exist..
PASS testObj2.testObject is defined.
PASS testObj2.testObject threw exception ReferenceError: Property testObject does not exist..
PASS successfullyParsed is true

TEST COMPLETE
